+++ mod.ts
@@ -1,12 +1,24 @@
 import * as commands from "./src/commands.ts";
 import * as expressions from "./src/expressions.ts";
 
+export { generateTypeScriptFromAction } from "./src/actions/generate_typescript.ts";
 export { defineWorkflows, type DefineWorkflowsProps } from "./src/config.ts";
 export { generate } from "./src/generate.ts";
 export { type AnyJob, Job, job } from "./src/job.ts";
 export * as Meta from "./src/meta.ts";
-export { type AnyStep, Step, step } from "./src/step.ts";
-export { type GetTemplate, Runner, Shell } from "./src/types.ts";
+export {
+  type AnyStep,
+  type DefaultStepInputs,
+  Step,
+  step,
+} from "./src/step.ts";
+export {
+  type ActionTemplate,
+  type GetTemplate,
+  Runner,
+  Shell,
+  type WithContext,
+} from "./src/types.ts";
 export { type AnyWorkflow, Workflow, workflow } from "./src/workflow.ts";
 
 export { commands, commands as c, expressions, expressions as e };