+++ changelog.md
@@ -4,6 +4,47 @@
 
 > [Compare](https://github.com/ifiokjr/actionify/compare/0.1.0...HEAD)
 
+### 🎉 Features
+
+- BREAKING! rename commands to align with `octokit` command names.
+  - `setMask` -> `setSecret`
+  - `setEnv` -> `exportVariable`
+  - `setError` -> `error`
+  - `setNotice` -> `notice`
+  - `setWarning` -> `warning`
+
+- Add a remote actions server to allow importing fully typed actions from `https://act.deno.dev`.
+
+  ```ts
+  import checkout from "https://act.deno.dev/actions/checkout@3.0.2";
+  import {
+    defineWorkflows,
+    e,
+    workflow,
+  } from "https://deno.land/x/actionify@0.1.0/mod.ts";
+
+  const checkoutStep = checkout((ctx) => ({
+    repository: e.wrap(ctx.github.repository),
+    ref: e.wrap(ctx.github.ref),
+    token: e.wrap(ctx.github.token),
+    lfs: true,
+  })).env((ctx) => ({
+    GITHUB_TOKEN: e.wrap(ctx.secrets.GITHUB_TOKEN),
+  }));
+
+  const ci = workflow({ name: "ci" })
+    .on("push")
+    .job("Explore-GitHub-Actions", (job) => job.step(checkoutStep));
+
+  export default defineWorkflows({
+    workflows: [ci],
+  });
+  ```
+
+### 🐛 Bug Fixes
+
+- Remove `@octokit/core` dependency which was adding `9mb` to the bundle size. Now only import types.
+
 ## 0.1.0
 
 > [2022-09-12](https://github.com/ifiokjr/actionify/compare/3d33388...0.1.0)